Bugfix Arduino code

+++ 3_Programação_RPi/3.5_SPI/Arduino_SPI_code/Arduino_SPI_code.ino
@@ -6,7 +6,7 @@ volatile char receivedData = 0;
 
 void blink(int led_pin, int num_blinks, int period)
 {
-  int t = millis();
+  unsigned long t = millis();
   period /= 2;
   while(num_blinks--)
   {
